refactor: simplify import statement
