+++ src/@types/index.d.ts
@@ -1,0 +1,11 @@
+export * from "./config";
+export * from "./event";
+export * from "./fonts";
+export * from "./format.formatted";
+export * from "./format.legacy";
+export * from "./format.owner";
+export * from "./format.v1";
+export * from "./IComment";
+export * from "./IPlugins";
+export * from "./options";
+export * from "./types";
