+++ src/comments/index.ts
@@ -1,0 +1,3 @@
+export * from "./BaseComment";
+export * from "./FlashComment";
+export * from "./HTML5Comment";
