+++ src/definition/index.ts
@@ -1,0 +1,4 @@
+export * from "./colors";
+export * from "./config";
+export * from "./fonts";
+export * from "./initConfig";
