+++ src/errors/index.ts
@@ -1,0 +1,4 @@
+export * from "./CanvasRenderingContext2DError";
+export * from "./InvalidFormatError";
+export * from "./InvalidOptionError";
+export * from "./NotImplementedError";
