+++ src/main.ts
@@ -1,29 +1,31 @@
-import type { CommentEventHandlerMap } from "@/@types/event";
-import type { formattedComment } from "@/@types/format.formatted";
-import type { IComment } from "@/@types/IComment";
-import type { inputFormat, Options } from "@/@types/options";
 import type {
   collision,
   collisionItem,
   collisionPos,
+  CommentEventHandlerMap,
+  formattedComment,
+  IComment,
+  inputFormat,
+  Options,
   Timeline,
-} from "@/@types/types";
-import { FlashComment } from "@/comments/FlashComment";
-import { HTML5Comment } from "@/comments/HTML5Comment";
-import { resetImageCache } from "@/contexts/cache";
-import { resetNicoScripts } from "@/contexts/nicoscript";
-import { plugins, setPlugins } from "@/contexts/plugins";
+} from "@/@types/";
+import { FlashComment, HTML5Comment } from "@/comments/";
+import {
+  plugins,
+  resetImageCache,
+  resetNicoScripts,
+  setPlugins,
+} from "@/contexts/";
 import {
   config,
   defaultConfig,
   defaultOptions,
+  initConfig,
   options,
   setConfig,
   setOptions,
-} from "@/definition/config";
-import { initConfig } from "@/definition/initConfig";
-import { CanvasRenderingContext2DError } from "@/errors/CanvasRenderingContext2DError";
-import { InvalidOptionError } from "@/errors/InvalidOptionError";
+} from "@/definition/";
+import { CanvasRenderingContext2DError, InvalidOptionError } from "@/errors/";
 import { registerHandler, removeHandler, triggerHandler } from "@/eventHandler";
 import convert2formattedComment from "@/inputParser";
 import typeGuard from "@/typeGuard";
@@ -38,7 +40,7 @@ import {
   processFixedComment,
   processMovableComment,
 } from "@/util";
-import { isFlashComment } from "@/utils/comment";
+import { isFlashComment } from "@/utils/";
 
 let isDebug = false;
